+++ libs/extensions/angular/image-banner/src/image-banner.component.html
